requirements.txt
# LocalAgent CLI Requirements
# Updated for Typer-based modern CLI architecture

# Core CLI Framework
typer[all]>=0.16.0
rich>=13.6.0
click>=8.1.0  # Still needed for compatibility
pyyaml>=6.0
aiohttp>=3.9.0

# Interactive Prompts
inquirerpy>=0.3.4
questionary>=2.0.0  # Alternative prompt library

# File Operations
aiofiles>=23.0.0

# Configuration and Validation
pydantic>=2.0.0
keyring>=24.0.0

# Provider SDKs
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.3.0

# Security
cryptography>=41.0.0

# MCP servers
computer-control-mcp>=0.3.0

# Plugin System
setuptools>=68.0.0
importlib-metadata>=6.0.0

# Development and Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
pylint>=3.0.0
mypy>=1.5.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.0.0

# Performance Monitoring
psutil>=5.9.0

# Additional utilities
typing-extensions>=4.5.0